Refactor timeout modules
