+++ src/components/timeout-modal/timeout-modal.dom.js
@@ -4,12 +4,12 @@ async function modals() {
   const timeouts = [...document.querySelectorAll('.ons-js-timeout-modal')];
 
   if (timeouts.length) {
-    const Timeout = (await import('./timeout-modal')).default;
+    const TimeoutModal = (await import('./timeout-modal')).default;
 
     timeouts.forEach(context => {
       let url = context.getAttribute('data-server-session-expiry-endpoint');
       let time = context.getAttribute('data-server-session-expires-at');
-      new Timeout(context, url, time);
+      new TimeoutModal(context, url, time);
     });
   }
 }
