+++ src/components/timeout-panel/examples/warn/index.njk
@@ -1,15 +1,13 @@
 {% from "components/timeout-panel/_macro.njk" import onsTimeoutPanel %}
 {{ onsTimeoutPanel ({
         "id": "countdown",
-        "countdown": {
-            "countdownInSeconds": 120,
-            "minutesTextSingular":"minute",
-            "minutesTextPlural":"minutes",
-            "secondsTextSingular":"second",
-            "secondsTextPlural":"seconds",
-            "countdownText":"For security, your answers will only be available to view for another",
-            "nojsText": "For security, your answers will only be available to view for another 2 minutes",
-            "countdownExpiredText":"For security, you can no longer view your answers"
-        }
+        "redirectUrl": "#!",
+        "minutesTextSingular":"minute",
+        "minutesTextPlural":"minutes",
+        "secondsTextSingular":"second",
+        "secondsTextPlural":"seconds",
+        "countdownText":"For security, your answers will only be available to view for another",
+        "nojsText": "For security, your answers will only be available to view for another 2 minutes",
+        "redirectingText": "You are being signed out"
     })
 }}
